+++ src/types/Component.ts
@@ -358,7 +358,7 @@ type StaticValuesSelectData = {
     data: {
         values: { label: string; value: string }[];
     };
-    dataSrc: 'values';
+    dataSrc?: undefined | 'values';
 };
 
 type JsonValuesSelectData = {