+++ src/types/project/settings/integrations/index.ts
@@ -2,4 +2,4 @@ export * from './dataConnections';
 export * from './email';
 export * from './eSign';
 export * from './fileStorage';
-export * from './reCaptcha';
+export * from './captcha';